+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Rejecting the initial message of a conversation leads to the conversation being 
 - Support for customizing domains in SIP headers via `EmergencySpecifics`.
 - Ability to remove message listeners.
 - Ability to specify endpoint type for each conversation.
+- Ability to specify custom user agent headers
 - Interfaces for easily fetching SIP information like TO, FROM and ROUTE headers.
 ### Fixed
 - Conversation state only changes if the state value itself changes. If origin is different but state is the same, it won't trigger a state change.